Switch to not forcing split of request channel handlers being seperate from the listener decl
