+++ src/features/cluster/delete-dialog/main/clear-as-deleting-channel-listener.injectable.ts
@@ -2,13 +2,19 @@
  * Copyright (c) OpenLens Authors. All rights reserved.
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
-import { getRequestChannelListenerInjectable } from "../../../../common/utils/channel/request-channel-listener-injection-token";
+import clustersThatAreBeingDeletedInjectable from "../../../../main/cluster/are-being-deleted.injectable";
+import { getRequestChannelListenerInjectable } from "../../../../main/utils/channel/channel-listeners/listener-tokens";
 import { clearClusterAsDeletingChannel } from "../common/clear-as-deleting-channel";
-import clearClusterAsDeletingHandlerInjectable from "./clear-as-deleting-handler.injectable";
 
 const clearClusterAsDeletingChannelListenerInjectable = getRequestChannelListenerInjectable({
   channel: clearClusterAsDeletingChannel,
-  handlerInjectable: clearClusterAsDeletingHandlerInjectable,
+  handler: (di) => {
+    const clustersThatAreBeingDeleted = di.inject(clustersThatAreBeingDeletedInjectable);
+
+    return (clusterId) => {
+      clustersThatAreBeingDeleted.delete(clusterId);
+    };
+  },
 });
 
 export default clearClusterAsDeletingChannelListenerInjectable;
