+++ src/main/app-paths/app-paths-request-channel-listener.injectable.ts
@@ -2,13 +2,17 @@
  * Copyright (c) OpenLens Authors. All rights reserved.
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
-import { getRequestChannelListenerInjectable } from "../../common/utils/channel/request-channel-listener-injection-token";
 import { appPathsChannel } from "../../common/app-paths/app-paths-channel";
-import appPathsChannelHandlerInjectable from "./app-paths-channel-handler.injectable";
+import appPathsInjectable from "../../common/app-paths/app-paths.injectable";
+import { getRequestChannelListenerInjectable } from "../utils/channel/channel-listeners/listener-tokens";
 
 const appPathsRequestChannelListenerInjectable = getRequestChannelListenerInjectable({
   channel: appPathsChannel,
-  handlerInjectable: appPathsChannelHandlerInjectable,
+  handler: (di) => {
+    const appPaths = di.inject(appPathsInjectable);
+
+    return () => appPaths;
+  },
 });
 
 export default appPathsRequestChannelListenerInjectable;
