+++ src/main/helm/repositories/get-active-helm-repositories/get-active-helm-repositories-channel-listener.injectable.ts
@@ -2,13 +2,13 @@
  * Copyright (c) OpenLens Authors. All rights reserved.
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
-import { getRequestChannelListenerInjectable } from "../../../../common/utils/channel/request-channel-listener-injection-token";
 import { getActiveHelmRepositoriesChannel } from "../../../../common/helm/get-active-helm-repositories-channel";
+import { getRequestChannelListenerInjectable } from "../../../utils/channel/channel-listeners/listener-tokens";
 import getActiveHelmRepositoriesInjectable from "./get-active-helm-repositories.injectable";
 
 const getActiveHelmRepositoriesChannelListenerInjectable = getRequestChannelListenerInjectable({
   channel: getActiveHelmRepositoriesChannel,
-  handlerInjectable: getActiveHelmRepositoriesInjectable,
+  handler: (di) => di.inject(getActiveHelmRepositoriesInjectable),
 });
 
 export default getActiveHelmRepositoriesChannelListenerInjectable;
