+++ src/main/helm/repositories/remove-helm-repository/remove-helm-repository-channel-listener.injectable.ts
@@ -2,13 +2,13 @@
  * Copyright (c) OpenLens Authors. All rights reserved.
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
-import { getRequestChannelListenerInjectable } from "../../../../common/utils/channel/request-channel-listener-injection-token";
 import removeHelmRepositoryInjectable from "./remove-helm-repository.injectable";
 import { removeHelmRepositoryChannel } from "../../../../common/helm/remove-helm-repository-channel";
+import { getRequestChannelListenerInjectable } from "../../../utils/channel/channel-listeners/listener-tokens";
 
 const removeHelmRepositoryChannelListenerInjectable = getRequestChannelListenerInjectable({
   channel: removeHelmRepositoryChannel,
-  handlerInjectable: removeHelmRepositoryInjectable,
+  handler: (di) => di.inject(removeHelmRepositoryInjectable),
 });
 
 export default removeHelmRepositoryChannelListenerInjectable;
