+++ src/main/utils/resolve-system-proxy/resolve-system-proxy-channel-responder.injectable.ts
@@ -3,12 +3,12 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { resolveSystemProxyChannel } from "../../../common/utils/resolve-system-proxy/resolve-system-proxy-channel";
+import { getRequestChannelListenerInjectable } from "../channel/channel-listeners/listener-tokens";
 import resolveSystemProxyInjectable from "./resolve-system-proxy.injectable";
-import { getRequestChannelListenerInjectable } from "../../../common/utils/channel/request-channel-listener-injection-token";
 
 const resolveSystemProxyChannelResponderInjectable = getRequestChannelListenerInjectable({
   channel: resolveSystemProxyChannel,
-  handlerInjectable: resolveSystemProxyInjectable,
+  handler: (di) => di.inject(resolveSystemProxyInjectable),
 });
 
 export default resolveSystemProxyChannelResponderInjectable;
