+++ src/main/utils/sync-box/sync-box-initial-value-channel-listener.injectable.ts
@@ -3,12 +3,19 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { syncBoxInitialValueChannel } from "../../../common/utils/sync-box/channels";
-import { getRequestChannelListenerInjectable } from "../../../common/utils/channel/request-channel-listener-injection-token";
-import syncBoxInitialValueChannelHandlerInjectable from "./sync-box-initial-value-handler.injectable";
+import { syncBoxInjectionToken } from "../../../common/utils/sync-box/sync-box-injection-token";
+import { getRequestChannelListenerInjectable } from "../channel/channel-listeners/listener-tokens";
 
 const syncBoxInitialValueChannelListenerInjectable = getRequestChannelListenerInjectable({
   channel: syncBoxInitialValueChannel,
-  handlerInjectable: syncBoxInitialValueChannelHandlerInjectable,
+  handler: (di) => {
+    const syncBoxes = di.injectMany(syncBoxInjectionToken);
+
+    return () => syncBoxes.map((box) => ({
+      id: box.id,
+      value: box.value.get(),
+    }));
+  },
 });
 
 export default syncBoxInitialValueChannelListenerInjectable;
